+++ beatbox/__init__.py
@@ -1,7 +1,6 @@
 from .vector import Vector
 from .raster import Raster
-from .do import Do
+from .do import Do, _build_kwargs_from_args
 from .convex_hulls import fuzzy_convex_hull
 # from .downloaders import HttpDownload # This is functional, but needs a re-factoring
-
 __all__ = ["Vector", "Raster", "Do"]
